requirements.txt
azure-functions
requests
openpyxl